requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.5
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==2023.10
cryptography==41.0.5
defusedxml==0.7.1
Django==4.2.7
django-allauth==0.58.2
django-axes==6.1.1
django-ckeditor==6.7.0
django-colorfield==0.10.1
django-crispy-forms==2.1
django-cuser==2017.3.16
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-filter==23.3
django-jalali-date==1.0.2
django-js-asset==2.1.0
django-mptt==0.15.0
django-phonenumber-field==7.2.0
django-tinymce==3.6.1
idna==3.4
jdatetime==4.1.1
kombu==5.3.3
modeltranslation==0.25
oauthlib==3.2.2
persiantools==3.0.1
phonenumbers==8.13.25
Pillow==10.1.0
prompt-toolkit==3.0.41
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.10
